+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 UNRELEASED
 - Various build file improvements
   https://gitlab.com/gitlab-org/gitaly/merge_requests/229
+- Implement FindAllCommits RPC
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/226
 
 v0.21.2
 - Config: do not start Gitaly without at least one storage